PNG-138 Small refactoring.
