+++ lib/utils/position.js
@@ -1,6 +1,6 @@
 import {isNum, int} from './shims'
 import ReactDOM from 'react-dom'
-import {getTouch, innerWidth, innerHeight, offsetXYFromParent, outerWidth, outerHeight} from './domFns'
+import {getTouch, innerWidth, innerHeight, offsetXYFromParent, outerWidth, outerHeight} from './dom'
 
 export function getBoundPosition(draggable, x, y) {
     // If no bounds, short-circuit and move on
